Add home button to toolbar, add back and bookmarks to menu
Also remove "Request desktop site" from menu since
I use desktop user agent by default for tablet builds
Make static fields of WSF$Site package-private

This makes a little more sense since they are accessed outside of the
class that they are defined in.

In addition, this fixes errors created by proguard when building with
OpenJDK.

Change-Id: I41d9bbb687628140038eb517ea08517cee37e0c7
Browser : Open in incognito tab

Long click on a link now allows you to choose
between opening in incognito/regular tab.

Thank Martin for the bright idea :)

Change-Id: Iba8e7446e69ae66e0bd957910e6309fdb0e0b947
Update Simplified Chinese translations for Browser

Change-Id: If6eb7d2af35ff9c1dd2aadf1af2720ecb253968b
move the User Agent selection out of debug

This helps the users on Tablets. Initially the UAString selection is
only possible when the Browser is built with debug enabled, by this
change, it is moved to advanced selection and available in the release
version.

Change-Id: If31e1e58be92e6b2e40ddf29bc98d4c4324f727d